% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_tbl}
\alias{area_tbl}
\title{Area Table}
\format{
## `area_tbl`
\describe{
  \item{division_name}{Division name, with four values}
  \item{area_name}{Area name, with 17 values}
  \item{lga_name}{Local government area name, with 79 values}
  \item{cent_long}{Longitude of centroid}
  \item{cent_lat}{Latitude of centroid}
  ...
}
}
\source{
<https://www.dffh.vic.gov.au/our-structure>
}
\usage{
area_tbl
}
\description{
A tibble of with 3 columns and 79 rows.
}
\keyword{datasets}
