% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dffh_area.R
\name{clean_dffh_area}
\alias{clean_dffh_area}
\title{Convert Victorian DFFH area names into a consistent format}
\usage{
clean_dffh_area(
  x,
  to = "dffh_name",
  fuzzy_match = TRUE,
  max_dist = 0.4,
  method = "jw"
)
}
\arguments{
\item{x}{a (character) vector containing Victorian DFFH area names or abbreviations.}

\item{to}{what form should the state names be converted to? Options are
"dffh_name", "dffh_name" (the default), and "dffh_code".}

\item{fuzzy_match}{logical; either TRUE (the default) which indicates that
approximate/fuzzy string matching should be used, or FALSE which indicates that
only exact matches should be used.}

\item{max_dist}{numeric, sets the maximum acceptable distance between your
string and the matched string. Default is 0.4. Only relevant when fuzzy_match is TRUE.}

\item{method}{the method used for approximate/fuzzy string matching. Default
is "jw", the Jaro-Winker distance; see `??stringdist-metrics` for more options.
#'}
}
\value{
a character vector of state names, abbreviations, or codes.
}
\description{
Convert Victorian DFFH area names into a consistent format
}
\examples{

library(vicdffhmap)

x <- c("Melburn", "Wang", "Donga")

# Convert the above to area names
clean_dffh_area(x)

# Disable fuzzy matching; you'll get NAs unless exact matches can be found

clean_dffh_area(x, fuzzy_match = FALSE)

# You can use clean_dffh_area in a dplyr mutate call

x_df <- data.frame(area = x, stringsAsFactors = FALSE)

\dontrun{x_df |> mutate(dffh_name = clean_dffh_area(area))}

}
