% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{division_map}
\alias{division_map}
\title{Division Maps}
\format{
## `division_map`
\describe{
  \item{division_name}{division name}
  \item{geometry}{geometry}
  \item{cent_long}{Longitude of centroid}
  \item{cent_lat}{Latitude of centroid}
  ...
}
}
\source{
<https://www.vic.gov.au/Divisional-model-department-education>
}
\usage{
division_map
}
\description{
Simple feature collection with 4 features and 4 fields
}
\keyword{datasets}
