% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dffh_area.R
\name{name_string_tidy}
\alias{name_string_tidy}
\title{Helper function used by clean_dffh_name}
\usage{
name_string_tidy(string)
}
\arguments{
\item{string}{a character string.}
}
\description{
Helper function used by clean_dffh_name
}
